Les modifications pour nanoplot et reads on conti pour les short_reads

+++ sources/assembly/metaflye_wraper.sh
@@ -5,6 +5,7 @@
 run="$1"
 output_directory="$2"
 technology="$3"
+cleanup="${4:-yes}"
 
 mkdir -p "$output_directory"/tmp
 
@@ -18,7 +19,7 @@ fi
 flye --meta -t $(nproc) --out-dir "$output_directory"/tmp/ $READ_TYPE "$run"
 
 mv "$output_directory"/tmp/assembly.fasta "$output_directory"
-if [ -f "$output_directory/assembly.fasta" ]; then
+if [ "$cleanup" != "no" ] && [ -f "$output_directory/assembly.fasta" ]; then
     rm -rf "$output_directory"/tmp/
 fi
